working on getting balances from API

+++ src/main/java/com/grmkris/mardukadmin/api/Balance.java
@@ -4,10 +4,11 @@ import lombok.Builder;
 import lombok.Getter;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Getter
 @Builder
 public class Balance {
     private String walletName;
-    private BigDecimal value;
+    private BigInteger value;
 }
